--- tests/fail-typecheck/fail_term_union.egg
@@ -1,13 +1,0 @@
-; Test that union on :nounion sorts fails
-; This test SHOULD fail because :nounion marks the sort as non-unionable
-; (This file is kept for manual testing - the fail-typecheck version is used by the test suite)
-
-(sort TermSort :nounion)
-(constructor term-fn (i64 i64) TermSort)
-
-; Create two terms
-(term-fn 1 2)
-(term-fn 3 4)
-
-; This should fail with: "Cannot union values of sort TermSort because it is marked as non-unionable"
-(union (term-fn 1 2) (term-fn 3 4))
