--- tests/term_annotation.egg
@@ -1,29 +1,0 @@
-; Test the :nounion annotation on sorts
-; :nounion marks a sort as non-unionable, meaning values of that sort cannot be unioned
-
-; Create a sort with :nounion and a constructor
-(sort TermSort :nounion)
-(constructor term-fn (i64 i64) TermSort)
-
-; Create some terms - this should succeed
-(term-fn 1 2)
-(term-fn 3 4)
-(term-fn 5 6)
-
-; Verify all rows exist as separate entries (not unioned)
-(check (term-fn 1 2))
-(check (term-fn 3 4))
-(check (term-fn 5 6))
-
-; Test :nounion with constructor :cost option
-(sort TermSort2 :nounion)
-(constructor term-fn2 (i64) TermSort2 :cost 5)
-(term-fn2 100)
-(check (term-fn2 100))
-
-; Test :nounion with constructor :unextractable option  
-(sort TermSort3 :nounion)
-(constructor term-fn3 (i64) TermSort3 :unextractable)
-(term-fn3 200)
-(check (term-fn3 200))
-
