--- tests/term_extraction.egg
@@ -1,17 +1,0 @@
-; Test :nounion sorts with extraction
-; :nounion sorts can still be extracted
-
-(sort TermSort :nounion)
-(constructor term-leaf (i64) TermSort)
-(constructor term-branch (TermSort TermSort) TermSort :cost 2)
-
-; Build a simple tree
-(term-branch 
-  (term-leaf 1)
-  (term-branch
-    (term-leaf 2)
-    (term-leaf 3)))
-
-; Extract should work on :nounion sorts
-(extract (term-leaf 1))
-(extract (term-branch (term-leaf 1) (term-branch (term-leaf 2) (term-leaf 3))))
