--- tests/term_with_rules.egg
@@ -1,29 +1,0 @@
-; Test :nounion sorts with rules
-; Rules can still pattern match on :nounion sorts and create new terms
-
-(sort TermSort :nounion)
-(constructor term-node (i64) TermSort)
-(constructor term-pair (TermSort TermSort) TermSort)
-
-; Create some term nodes
-(term-node 1)
-(term-node 2)
-(term-node 3)
-
-; Create some pairs
-(term-pair (term-node 1) (term-node 2))
-(term-pair (term-node 2) (term-node 3))
-
-; Rule to propagate terms - this creates new term entries but doesn't union
-(rule ((term-pair x y))
-      ((term-pair y x)))  ; Create reverse pair
-
-(run 1)
-
-; Check that the original pairs exist
-(check (term-pair (term-node 1) (term-node 2)))
-(check (term-pair (term-node 2) (term-node 3)))
-
-; Check that the reverse pairs were created
-(check (term-pair (term-node 2) (term-node 1)))
-(check (term-pair (term-node 3) (term-node 2)))
